+++ gapipy/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '2.14.6'
+__version__ = '2.15.0'
 __title__ = 'gapipy'