+++ gapipy/resources/geo/country.py
@@ -1,7 +1,8 @@
 # Python 2 and 3
 from __future__ import unicode_literals
 
-from ..base import Resource
+from gapipy.resources.base import Resource
+
 from .state import State
 
 
@@ -10,5 +11,12 @@ class Country(Resource):
     _resource_name = 'countries'
     _is_parent_resource = True
 
-    _as_is_fields = ['id', 'href', 'name']
-    _resource_collection_fields = [('states', State)]
+    _as_is_fields = [
+        'id',
+        'href',
+        'name',
+    ]
+
+    _resource_collection_fields = [
+        ('states', State),
+    ]